ui improvements
